refactor: 희망 직무 user 패키지로 이동
