+++ core-application/src/main/java/gohigher/user/port/in/DesiredPositionRequest.java
@@ -1,4 +1,4 @@
-package gohigher.position.port.in;
+package gohigher.user.port.in;
 
 import java.util.List;
