+++ in-adapter-api/src/main/java/gohigher/user/DesiredPositionCommandController.java
@@ -1,4 +1,4 @@
-package gohigher.position;
+package gohigher.user;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -6,21 +6,21 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import gohigher.auth.support.Login;
-import gohigher.position.port.in.DesiredPositionRequest;
-import gohigher.position.port.in.PositionCommandPort;
+import gohigher.user.port.in.DesiredPositionCommandPort;
+import gohigher.user.port.in.DesiredPositionRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequiredArgsConstructor
-public class PositionCommandController implements PositionCommandControllerDocs {
+public class DesiredPositionCommandController implements DesiredPositionCommandControllerDocs {
 
-	private final PositionCommandPort positionCommandPort;
+	private final DesiredPositionCommandPort desiredPositionCommandPort;
 
 	@PostMapping("/v1/desired-positions")
 	public ResponseEntity<Void> saveDesiredPositions(@Login Long userId,
 		@RequestBody @Valid DesiredPositionRequest request) {
-		positionCommandPort.saveDesiredPositions(userId, request.getPositionIds());
+		desiredPositionCommandPort.saveDesiredPositions(userId, request.getPositionIds());
 		return ResponseEntity.noContent().build();
 	}
 }
